requirements.txt
Flask
numpy
Pillow
tensorflow
scikit-learn
gunicorn